(Status_Code_Image): new type to handle 3 digits status code (Status_Data): Code is now a Status_Code_Image (Status_Messages): change declaration to match new design.
*** From Sune Falck:
Changed Image to handle Natural instead of Positive, gave
constraint_error for 0 in seconds and minutes. Image now returns a
leading zero for numbers with only one digit so that time and day
representation will be correct. Changed Month_Name to an array
instead of a function. Changed Month_Number to use the array
Month_Name. Added function Weekday.