* aws-net.adb: (Add) [Set_Access]: Removed. Not needed anymore. (Adjust): New routine for implementing a ref counter on FD_Set. (Finalize): Idem. (Initialize): Idem. (Free) [FD_Set]: Default implementation, does nothing.
* aws-net.ads:
(FD_Set): Remove the Size discriminant. Not used in this abstract
 interface. This type is now controlled for automatic memory
 deallocation.
(Reallocate): Now a procedure instead of a function. The reallocation
 is done directly into the FD_Set passed as parameter as it is an
 access type. This is cleaner than the previous implementation
 returning an access to the new FD_Set and was for this exposing
 the access.
(Size): New routine, needed as the FD_Set does not have a Size
 discriminant anymore.
(Free) [FD_Set]: Default implementation, does nothing.

* aws-net-poll_events.ads:
(Set): Discriminant is now Initial_Size instead of Size. Required as
 it is not possible to have default parameter for a tagged record so
 we cannot change the Size discriminant during the life of the object.
 The real size is not a field of the record. Fds field is now an
 access type to be able to change dynamically its size.
(Size): New routine.

* aws-net-poll_events.adb:
(Add): Add the check for reallocation here. Was in the Add routine
 taking an access type as parameter.
(Free) [FD_Set]: Free the Fds set.
(Reallocate): New version. It is now a procedure instead of a function
 returning the reallocated object.
(Size): New inherited routine. Returns the size of the array.
(Wait): Adjust call to Thin.Poll to properly pass the address of
 the Fds.

* aws-net-generic_sets.adb:
(Add_Private): Slight adjusment to the Add call. The version with an
 access type does not exists anymore.

* aws-net-generic_sets.ads:
(Socket_Set_Type): Change Set_Access to FD_Set_Access.
--
Another batch of clean-up for the keep-alive implementation.
This new implementation propose a better interface for the FD_Set.