Add a simple SMTP server API.
With this API it is possible to get the message sent by
a client. The first purpose is to be able to test the
client API into the regtests. This API can also be used
to build a mail server.

File a test case for the SMTP support.
--
Work for H220-022.