Rename discriminant SOAP to XMLADA and used it to jabber tests.
The discriminant for the tests was named SOAP because SOAP was
using XML/Ada in fact. But Jabber needed also a check, it is
cleaner to use XMLADA as a discriminant as this is really what
is needed to support those tests.
Fix Jabber test when XML/Ada is not set.
--
Work for H303-007.