Ensure that ADA_PROJECT_PATH is set dynamically.
To set ADA_PROJECT_PATH the environment variable $PWD was used.
This variable is set by the underlying shell. But if an application
is launched which cd into the AWS root directory to issue the make
command, then $PWD is wrong as it point to the directory from where
the script is run and not the current working directory. Using
$(shell pwd) instead fix this problem as it computes the current
working directory dynamically.
--
Work for H222-024.