Remove Replace as the Update routine is equivalent (Tables.set).
We already add an Update routine to either add or replace a
parameter association. The newly introduced Replace routine
is not needed and removed. At the same time the Replace
routine added in Parameters.Set has been renamed Update for
consistency. The non regression test is updated accordingly.
--
Code clean-up.