Fix handling of attachments when passing argument to POST messages.
AWS was not parsing properly POST requests using the AWS.Client API
with attachments and with some parameters in the message body. This
problem resulted in attachments lost or POST parameters not present.

Add corresponding regression test.

Fix some test output to match proper attachment names. This side
issue has been found and fixed with the rework of the attachment
support.

Work for HA21-048.