Fix SOAP digest authentication.
SOAP request are sent in streaming mode. In this case the body is
not read just after parsing the headers. In case of digest authentication
it is important to read the body is response 401 is retrieved otherwise
some garbadges are let on the stream which cause a protocol error.
Add regression test.
--
Fixes I308-003.