Enhance wsdl2aws to generate stubs based on connection.
New stubs using an HTTP_Connection are now generated by wsdl2aws. It
is then possible to reuse a connection for multiple SOAP requests.
This brings speed improvement when many SOAP calls are to be done on
the same server. The Endpoint based routines are now calling the
connection based ones (no code duplication).
A regression test is added.
--
Work for I309-020.