Rework the way SOAP digest authentication is fixed.
In fact we now make sure the body has always been read before issuing
another request attempt to the server. This is important otherwise
instead of receiving the response to the new request we would parse
body of the previous rejected message (authentication failure is one
case).
--
Better fix for I308-003.