Specify full path for projects.
Projects used for the installation and to build AWS are
separate. We can specify the full path to the generated
projects in .build/projects instead of requiring that the
ADA_PROJECT_PATH be set.

Also now all the tools and demos are built in static mode only.
We do not need to set the PATH to find the DLLs in the makefile.

General makefile clean-up.