Generate new routine Socket_Errno.
For portablility issue it is better to avoid use of GNAT.Sockets.Thin.
The last socket errno is using the standard errno on UNIXes but on
Windows a specific routine is to be called. The Socket_Errno routine
here hide this OS dependency and will make it possible to avoid any
use of GNAT.Sockets.Thin.
--
Part of work for I319-025.