Add support for adding Ajax action in table.
To be able to add a set of Ajax actions into a TABLE
statement we need to be able to specify unique JavaScript
id/name. For this a new parameter has been added to be
able to pass an ID_INDEX (as parameter 18). For example
this parameter can be passed the @_TABLE_LINE_@ tag.