Raise constraint_error when no data read and D'First - 1 out of bound.
When no data is read the Last index was set to Data'First - 1 or
Data'Last when Data'First = Stream_Element_Offset'First. To comply with
recommendation in AI95-227 and new GNAT implementation for GNAT.Sockets
for example a Constraint_Error is now raised in this later case.

This change has been done for consistency reasons as the GNAT runtime
now follows the recommendation on AI95-227.

A common case to have Data'First = Stream_Element_Offset'First is:

   I  : Stream_Element;
   A  : Stream_Element_Array;

   I & A;

This is easily worked around by using:

   (0 => I) & A;

--
Work for I915-011.