New implementation of the Web_Block context support.
This new implementation separate the context from the session handling.
This new context is now content based. A context id depends on the
content in the context. Each page generated will have a different context
id if the context has been updated.

This fix a serious problem with the browser's tab or back button which
could corrupt an application by using a wrong context.
--
Work for I804-012.