Fix Web_Block AWS test 0160.
To be able to run this test AWS Web_Block framework has been
enhanced to be able to inject the context into a specific
location (tag) placed into the template. This feature could
be useful in other contexts.

Fix 0160_wb_test to be able to pass with the context support.
--
Continued work for I804-012.