Make poll support generic.
We want to be able to use the poll emulation, using select OS call,
on other OS than Windows. This will be used on VxWorks for example.
This patch removes Win32 specific code and use a small C wrapper
to be able to deal with fd set (using C macros).
--
Work for IC11-015.