Reorganize sources under src directory.
The first goal is to have the AWS sources under the same directory
for clarity. The second goal is to separate the core sources from
auxiliaries (optional to have a running server). This work will
be useful for small embedded server. Motivated by the work on
VxWorks port.
--
Motivated by IC11-015.