Generate global targetNamespace in schema when possible.
This patch is to check if all schema definitions are having
the same namespace. If so, we place a targetNamespace node
into the main schema node. This is to work around a Microsoft
.Net toolset bug where targetNamespace into schema's type
definitions are not taken into account. This is of course
mandatory to avoid name clashes (in Ada two types with the
same name in different packages for example).
--
For IC10-001.