Properly handles multiple parts responses in wsdl2aws.
A response of a SOAP routine having multiple parts is handled as a record.
In this case though there is no name for this record type nor associated
namespace. This patch fix this issue by using the AWS default namespace
in this case. The SOAP record type name is derived from the name.
--
Work for J217-009.