Do not check for OpenSSL libraries on UNIX.
This does not work on 64 bits machines where the 32 bits versions
of the libraries are also installed. This does not cover GNU/TLS
either. Anyway, assuming the OpenSSL libraries are installed in
the standard location, the linker will add the proper path
(32 vs 64 bits) on the command line.

So on UNIX we set the library path to "." (a directory that always
exists).
--
For IB13-008.