Fix test_zlib by using a simple random generator.
To ensure a stable output for this test a simple random generator
as been introduced. We want to avoid using the GNAT one that may
be improved from time to time. For this test it is not important
that the random numbers be good anyway.

The regression was introduced by E513-008.
--
Fixes J310-013.