Fix status code for Response.File recorded in log file.
The status code recorded into the log file was always 200 for file
resources. The 404 (not found) or 304 (up-to-date) was not properly
set in the log file. Note that the proper status was returned to the
web browser.

The status code is now properly updated in those cases when sending
file resources. This ensures that the correct status is recorded.

Add corresponding regression test.
--
For K307-010.