Use regexp instead of glob pattern for Register_Regexp.
The documentation was correct but the implementation of the Register_Regexp
in AWS.Services.Dispatchers.URI was wrong. A globing pattern was used
instead of a regular expression.

This patch correct this. The test 0048_dispatch has been enhanced to
check for regular expressions.

Add a note into the non upward compatibility section.
--
For K308-027.