Unconditionally free the resource.
When calling AWS.Response.Set.Clear on a Response.Data object we always
want to free the stream access. Only closing must be done when the stream
has not be taken by the user.

This fix an issue when a message body is retrieved via a Resources.File_Type
with a persistent connection and reusing the same Response.Data object.

Add corresponding regression test.

For K615-002.