Fix memory leak in Client.HTTP_Utils.Disconnect.
If a connect is done on a domain which does not exit, the socket
will be allocated in Connect, but Connection.Opened will still be
False. In this case, we do not want to Shutdown the socket but we do
want to free the allocation for the socket to avoid a memory leak.

Path contributed by Wiljan Derks.

Add corresponding regression test.

For KA08-001.