Don't assume pdf documents get generated
None of the BSDs have pdf generation capability by default.  It
may be available as part of a huge package optionally from the
package repository, but more often than not, the pre-requisites
for generating pdfs will not be present.

This patch will check for the presences of the aws.pdf, and if it
doesn't exist, no attempt will be made to copy over the aws.pdf
nor the templates_parser.pdf file.  This will avoid two nasty
"file not found" messages in the middle of the installation step.