Fix code to avoid active waiting.
The Ready entry must return only when there is socket presents
(when a download is ready to be started) and not when there is downloads
waiting in the queue. Ready to be started here means that the client
is actually waiting for some data. This avoids having an empty socket
set on which a wait returns immediately as there is nothing to wait
for.

Fixes KA10-018.