Use existing AWS API to handle cookie.
The Get and Exists functions searched the cookie string in
a very weak and error prone manner. This has been updated
to instead use some of the available AWS tools so it's now
much more reliable.

Also did some minor tweaks to the Get functions that convert
string values to Integer and Float.