When maximum number of POST parameters reached, return 403 error code.
We were previously returning a 500 error code (Internal Server Error).
This is wrong, we have detected a possible security issue and this is
not a server error but more a client error so a 4xx error code seems
better. So the 403 (Forbidden) error code is now returned.

Continued work for L120-005.