Fix posix implementation of poll.
The previous implementation was wrongly expecting that the fd of
value n would be stored into the nth bucket of the fd set. The fd of
value one is in fact stored into the lower bit of the first word (more
precisely the first fd_mask) of the set.

Fixes L329-055.