Fix Do_Handshake_Internal in multi-theaded application context.
This fix is still misterious. We need to yield in the case of
a server before callback again gnutls_handshake(). Not doing so
will just crash the application.

After some analysis it seems that without yielding a lock is
not created (premption fails?) and this in turns crash the
application.

This behavior can be reproduced with the 0226_client_cert regression
test.