Add proper support for cross-environments.
This patch makes it possible to build AWS in native and cross
environments from the same tree. That is the build happen now
into .build/<target_name> AWS's root directory. The installation
is also done in specific directories for the cross builds
$prefix/$target.

Moreover generating OS constants and type definitions (in AWS.OS_Lib)
is done using xoscons and a template. This way it is possible to
create AWS.OS_Lib for a cross-target. It is then now possible to
remove the VxWorks specific AWS.OS_Lib (aws-os_lib__vxworks.ads) as
this can now be generated.

Generally this patch fix wrong uses of host kind where target is
meant. Consequently, some clean-ups and simplifications have
occured in makefiles.

For LA06-001.