WebSocket code refactoring.
The protocol specific state is now on a protocol specific record and
not part of the global WebSocket object. This is cleaner and will
make it easier to add support for other protocols if necessary.

It also make it clear which information is used in a given protocol.

Motivated by M527-001.