Add On_Message version with an Unbounded_String.
This version is to be used to handle large messages. Note that the default
implementation of this callback is to call the On_Message with a string. So
it is upward compatible.

An end-user has the choice to override this version to handle large messages
or the version with a string when it is known that only simple messages are
to be handled.

Continued work for M527-001.