Rework the closing behavior for the WebSocket protocol.
We now detect properly wrong closing error code and fail when the
frame is larger than 125 bytes (as for all control frames).

This fixes the Autobahn tests in the category 7. At this stage all
tests not dealing with UTF-8 messages are passing. It means that only
the tests in category 6 and 7.5.1 are failing.

Continued work for M527-001.