Add timeout for WebSocket Send and Close.
With the client side API we want to have a way to specify a timeout to avoid
blocking if a client is blocked or not responding. For this the Send and
Close routines have now a Timeout parameter. The default value is Forever
which means blocking and corresponds to the previous behavior.

For M723-005.