Ensure that Socket_Write do write all data into the buffer.
We are not sure that this routine will be called again (for example
last chunk of a large data), so we want the Socket_Write to do a
blocking send and send all data into the buffer. Not doing that
we are missing last chunk of data on slow connection which might
be saturated and won't accept all data.

For M802-001.