Add Send routine for a single socket in the registry.
It was possible to send a message to a set of recipients (sockets
matching a specific URI and/or Origin). It is now possible to
send a message to a single socket from its handle. Those routines
are equivalents to the ones in AWS.Net.WebSocket but are threading
safe and can be mixed with other incoming or outgoing messages.

Part of MA14-060.