Properly setup a concurrent access for the Web_Object map.
It is wrong to access any Ada constainers concurrently (even for
reading). The solution here is to wrap the routines used to access
the containers into a protected object. Note that all protected
routines are procedure to ensure exclusive access to the protected
ressource.

This fixes a regression on the test suite when finalizing the container.