Add support for controlling WebSocket handshake.
A new unit AWS.Net.WebSocket.Handshake_Error is introduced.

In the user's WebSocket factory it is now possible to return a WebSocket
object from AWS.Net.WebSocket.Handshake_Error to reject the handshake.
With such response it is possible to specify a specific status code
and an associated reaso-phrase. The default status code is 403.

Furthermore, if the user's factory raises an exception the WebSocket
handshake is rejected with a 403 HTTP error code.

For N227-068.