Rework the support for MIME_Types configuration.
This routine was per-server but this was unfortunate as the MIME types
table is anyway global to the process. So now, the MIME_Type is a
per-process option.

Also, we want to load the MIME_Types specified as configuration in the
user's code when starting the server.

Add corresponding regression test.

For N731-023.