Some Jabber enhancements.
Add bi-directional support (to process received messages),
the ability to subscribe and unsubscribe/remove Jabber IDs from
the account's roster, and the ability to use the same presence
and message callbacks with multiple accounts while maintaining
the context of which account the message came from.

The other change made was to add a serial number to use with
message IDs to ensure that they were unique.

Contributed by Jesse Lang.

For N113-043.