Keep the name-space information when parsing types.
This fix issues with derived type where the base type name-space was set
to the derived type name-space. This is not always correct.

This made it possible to fix the character handling. It is now not
necessary that the character type be of name Character. The parser now
automatically detect a simpleType corresponding to a character whatever
the name given to the type. If the simpleType does not correspond to a
character it just generate the necessary code to handle this type as for
all other schema defined types.

Add corresponding regression test.

Part of O118-004.