Add support for signed server's key file.
It is now possible to use a signed key file with AWS. In this case
a pass-phrase is requested to be able to start the HTTPS
session. This pass-phrase can be handed over to the server using
the Set_Password_Callback in AWS.Net.SSL.Certificate.

The HTTPS demo has been enhanced to demonstrate this.

Add corresponding regression test.

For 0202-018.