Initial version of the WSDL/SOAP document style binding.
This is the initial support for the document style binding in WSDL and
in SOAP messages. For handling SOAP messages with the document style binding
at runtime a lot of information from the WSDL is needed.

This patch introduce a shema object which is used in many places to
keep record of the types (from the schema) definition. It also keep
record of the actual binding (document or rpc) and the encoding of
the requests and response.

To call a document operation the signature of the call is needed. The
signature (set of parameter names) of every operations from the WSDL
are also recorded together with the actual SOAP operation.

Also, in many places the name-space is now preserved. Instead of
defaulting to the standard xsd type we can now handle properly the
user's definitions are described in the schema.

Add corresponding regression test.

For O421-035.