Parse and record all standard name-space.
We do not to assume that xsd, xsi, soapenc and soapenv are the prefix
used. These are the common names but it can be whatever the user decide
to use on the WSDL for example.

We now keep record of the actual name used and stored them into the
schema which is used consistently by all routines to check for the
proper prefix to use.

This fixes a number of regressions due to the new more conforming
SOAP circuitry.

Continued work for O421-035.