A document style SOAP message is identified by the paremater's type-names
The type name is streamed into the SOAP message instead of the parameter
name. This fixes an issue with the previous implementation. Also a SOAP
call is identified by the signature of the call using the parameter's
type-names and not the parameter's names.

Add corresponding regression test and adjust the expected output of
0271.

Continued work for O421-035.