Fix document/litteral SOAP message for complexType.
A complex type must be streamed using the name of the element and
not the type name. This patch properly register the element name
and set the enclosing element name where appropriate.

Also the case where a complexType is directly inside an element node
has been fixed.

Finaly the check for type using the schema has been fixed to properly
check for schema (not assuming xsd:) prefix as found in the schema.

Fix expected output of 0271 test.

Add corresponding regression test.

For O421-035.