Fix schema handling in WSDL.
First the tool ada2wsdl does not generate targetNamespace in elements
inside the schema as this is not valid. When there is multiple schema
used, ada2wsdl generates multiple schema one for each targetNamespace.

At the same time the wsdl2aws has been fixed to support multiple
schema definition.

This change was needed to properly support derived types for which
the base type are found in different packages. So have different
targetNamespace.

Some tests expected output have been changed due to the change above,
the new output are better as the original namespace is kept.

Add corresponding regression tests.

For Q309-017 and Q301-031.