Activate the GNAT option to detect unused entities.
This option was disabled by mistake (R vs r option). This has given the
oportunity to clean-up code by removing duplicate with/use clause and
also a generic routine not used anymore.

At the same time we now use -Wno-implicit-function-declaration for Zlib
C code which is an external set of sources synchronized with upstream
from time to time.

For Q426-015.