Fix SOAP timeInstant handling regarding time-zone.
This patch first makes a timeInstant always represented as a local time.
The subtype Local_Time makes this clear now. The converstion is done
when exchanging the time over the network. That is, the sent or received
time is always represented in UTC.

This fixes at the same time support for tiùe-zone which has an offset
of some minutes and not only full hours.

Update regression test accordingly.

For Q714-006, better fix for Q707-010.