Do not lock forever when a WebSocket is hard-disconnected.
When a WebSocket is not reacheable anymore because the network cable
has been disconnected we do not want to lock. In such a condition a
wait (poll system call) would wait forever. So in the Send routine
we now use a timeout. If a WebSocket is detected to be timed-out we
either unregister it or we call the new user's callback error for
the given WebSocket.

For RC13-003.