New implementation of ada2wsdl based on LaL.
This new implementation is based on LaL and do not use ASIS anymore.

The -I options has been removed as not supported anymore. The LaL parser
is project based exclusively.

Some expected output have been updated as the new implementation does
better work with the name-space by avoiding generating some NS that are
not used. So the NS prefix may be different.

We disable the check on routines ordering in ada2wsdl-parser.adb as
some generics make it impossible to follow this rule.

Note that this new implementation does far better job as computing the
range (range <low> .. <high>) and array indexes. The new parser walk up
the tree and do proper computation of the values instead of relying on
some simple format checked by regexp. So for example a range like:

   range 1 + 6 - 4 .. integer'last - 7

Will be properly computed avoiding some restriction on the specs.

For S415-009.