Rework AWS client's API to avoid stressing too much the stack.
When sending large payload over a socket with AWS.Client.Post we
ensure that no copy of the data is made on the stack to avoid a
stack overflow. This is especially important when a call is made
from a task where the stack may be limited.

Add corresponding regression test.

Fixes T527-014.