Config.Ini.Read: review documentation and fix implemenation.
All exceptions where masked which was not compliant with the comments.

Now this routine properly propagates the exception Constraint_Error in case
of erroneous option name or value.

The regression was introduced in 05a2aea when fixing a possible resource
leak failing to close the file descriptor of the configuration file in
case or error.

For T707-002.