Add support for optional values from server's SOAP messages.
When AWS receives a SOAP message with a component xsi:nil="true"
a xsd:null object was created but the deserialization was
raising an exception as not founding the proper SOAP type.

This has been fixed by returning a default value (0, 0.0, False or
empty string) for the corresponding Ada variable/record.

Add a corresponding regression test.

For U122-023.