Fix issue where From_Utf8 was raising an exception.
Some characters don't have an equivalence from Utf8 to 8bits. In
this case we fallback by default to a space character to avoid
the exception. A callback is also put in place to let user's propose
some equivalence better the Utf8 character and 8bit one. For example
in Utf8 there is different left and right apostrophe, one can
use the standard 8bit quote in place.

For U127-049.