Make sure that reading & writing from WebSocket is safe.
A WebSocket can be freed/released while reading or sending messages
to/from a WebSocket. To handle this cas we record a WebSocket Id on
the various containers and ensure that a WebSocket is still properly
registered before handling it. This also ensure a deferred free
of the WebSocket.

Continued work for U504-028.