Rework the max concurrent stream check.
A stream must be kept after it is closed as some more frames can be
received. In the current implementation we have no concurrency on
stream, each one is created, used an closed. So we cannot use a
bounded containers for the number of concurrent stream check.

This fixes some regressions now that the connection is not closed
after every stream.

Part of S507-051.