Fix 0043_check_mem test on GNUTLS-3.7.2
* regtests/0043_check_mem/check_mem.adb:
Use separate TLS 1.2 client connection to test sessions.

* regtests/0033_afile_sec/s_afile_pack.adb:
Improve to test for nonempty sessions.

* regtests/0043_check_mem/test.py:
Do not print gnatmem info difference, it is too big and not convinient to
analyse. Use switche --disable-cleanup to look at gnatmem info in the
temporary working directory.

* src/core/aws-net-ssl-certificate.ads:
Unitialized certificate equal to Undefined to be able to compare with
unitialised certificates.