Full rework of the SOAP generator.
The types, callbacks, main, client & server code is now generated
using templates files. This brings a large clean-up in the Ada
code of the generator and makes easier to change the actual
code generated.

Some refactoring has also been done to simplify the generated
code making it more generic. There is still some room for
improvements, but this is a big milestone.

Minor update of exepected output to cope with the differences due
to the new code generation layout.

For V531-036.