XFAIL a number of SSL-only testscases
We noticed when running the testsuite with SSL activated that
a number of testcases were failing. We will investigate those failures
under VC15-012, and in the meantime, to acknowledge that those failures
do not look critical, this commit is XFAIL-ing those tests.

In doing so, this commit adjusts a number of "!ssl DEAD" commands
in favor of using the "ssl REQUIRED" command instead, which has
the advantage of not being "cancellable" by other commands.
Without this, the order of the entries in the test.opt can be
treacherous.

TN: VC15-012