Make testcase endianness-independent
On ppc-linux, the output for this testcase is as shown in the new test.out.
The differing output comes from the testcase writing multi-byte integers to
an Ada stream: the way those integers are serialized into "stream elements"
(i.e. into a bytestring) is target-dependent by default.

This means that by default, on PPC, the zlib algorithm processes this hex
string…

    00 00 00 01  00 00 00 02  00 00 00 03  00 00 00 04

… instead of that hex string on x86_64…

    01 00 00 00  02 00 00 00  03 00 00 00  04 00 00 00

… and the latter happens to compress "better" than the former (yielding a
shorter compressed form).

The solution presented by the GNAT RM § "RM 13.13.2(1.6): Stream Oriented
Attributes" is to use a binder switch to request a consistent serialization
protocol (XDR)… which happens to be big-endian, so we need to update
test.out in the end.

TN: eng/toolchain/aws#3