Fix set parsing in SOAP messages.
A new routine Parse_Set is introduced and the is_a schema
for a set is recorded as set to ensure proper parsing as
it cannot be parsed as an array. A may not have a root
node, for example when it is to record element having
more than one occurrence (maxOccurs > 1).

TN eng/toolchain/aws#46