An unknown status code from server should be handled gracefully.
If a server send a status code which is not recognized by AWS the
class status code should be used. The class status code is given
but the first digit of the status code (471 should be handled as 400).

This patch introduce a new routine with the proper semantic and call
it when handling the HTTP/1.1 and HTTP/2 protocols.

TN eng/toolchain/aws#58