Setup CA certificates using OpenSSL or GNUTLS API.
Remove the setting for the default CA certificate file. This is Debian
specific and not working on Windows anyway.

The OpenSSL and GNUTLS implementation now use the proper API to load
the CA certificate chain.

Also refactor code for setting the CA certificate in OpenSSL avoiding
a large duplicated code.

TN eng/toolchain/aws#61