Only send headers for status code S404 (no found) or S304 (up-to-date).
This ensure that the content-length header is set to 0 which is
correct as there is no body in those two cases. It avoids client
to wait long time for the answer and finaly terminate because the
connection is beeing closed as not active.

Adjust the client part accordingly, if content-length is 0 we don't
have to read a body.

TN eng/toolchain/aws#65