SOAP_Attribute: New type derived from SOAP_Record.
This is a compound type used to record XML attributes. The first
item is called attribute_value and contains the value of the node
(named attribute). The other fields are to record the attribute's value.

Update demos and some regression tests to avoid ambiguous use
of contructor R.

TN eng/toolchain/aws#108